# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dictionary App for IT",
        "description": "API Docs for Dictionary",
        "contact": {
            "email": "quanghuybest@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/get-word-lookup-history/{user_id}": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Lấy từ vựng trong lịch sử tra từ theo id người dùng",
                "operationId": "d5360c1b4713a3777141804018287acd",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Nhập id của người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy thành công từ vựng trong lịch sử tra từ.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy người id dùng!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Lỗi rồi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/display-by-time-word-lookup-history": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Display by Time Word Lookup",
                "description": "Search for word lookup by user id and english",
                "operationId": "dc4a47a1727c01bb258f4b98656a403f",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "time",
                        "in": "query",
                        "description": "day/month/year",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "18/12/2023"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Không tìm thấy tài nguyên!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "validator_errors": {
                                            "type": "object",
                                            "example": {
                                                "time": [
                                                    "Vui lòng nhập ngày giờ"
                                                ],
                                                "user_id": [
                                                    "User id phải là số!"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/display-by-time-translate-history": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Display by Time Translate History",
                "description": "Search for translate by a user id and english",
                "operationId": "f014ab3600585c1f04c8858133d155f4",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "time",
                        "in": "query",
                        "description": "day/month/year",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "18/12/2023"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Không tìm thấy tài nguyên!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "validator_errors": {
                                            "type": "object",
                                            "example": {
                                                "time": [
                                                    "Vui lòng nhập ngày giờ"
                                                ],
                                                "user_id": [
                                                    "User id phải là số!"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/delete-by-id-word-lookup-history/{user_id}/{id}": {
            "delete": {
                "tags": [
                    "History"
                ],
                "summary": "Delete By Id Word Lookup History",
                "operationId": "c02d3d59a4517199be2c207513579828",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Word id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Không tìm thấy tài nguyên"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/delete-word-lookup-history/{user_id}": {
            "delete": {
                "tags": [
                    "History"
                ],
                "summary": "Delete All Word Lookup History",
                "operationId": "a3f471f1a39630df059c641cac1abfcb",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Không tìm thấy tài nguyên"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/get-translate-history/{user_id}": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Lấy bản dịch trong lịch sử dịch theo id người dùng",
                "operationId": "5ddbbf544a0d45ac8b8c5a90d4a4a456",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Nhập id của người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy thành công bản dịch trong lịch sử tra từ.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy người id dùng!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Lỗi rồi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}